refactor rasta types from udp header
