+++ src/rasta/headers/udp.h
@@ -12,7 +12,8 @@ extern "C" {  // only need to export C interface if
 
 #include <stdint.h>
 #include <netinet/in.h>
-#include <config.h>
+
+#include <types.h>
 
 #ifdef ENABLE_TLS
 #include <wolfssl/options.h>
@@ -20,41 +21,6 @@ extern "C" {  // only need to export C interface if
 #endif
 
 #define IPV4_STR_LEN 16
-
-#ifdef ENABLE_TLS
-enum RastaTLSConnectionState {
-    RASTA_TLS_CONNECTION_READY,
-    RASTA_TLS_CONNECTION_ESTABLISHED,
-    RASTA_TLS_CONNECTION_CLOSED
-};
-#endif
-
-struct RastaState
-{
-    int file_descriptor;
-    enum RastaTLSMode activeMode;
-    const struct RastaConfigTLS *tls_config;
-#ifdef ENABLE_TLS
-    WOLFSSL_CTX* ctx;
-
-    WOLFSSL* ssl;
-    enum RastaTLSConnectionState tls_state;
-
-#endif
-};
-
-// #if defined(ENABLE_TLS) && defined(ENABLE_TCP)
-#ifdef ENABLE_TLS
-    struct RastaConnectionState
-    {
-        const struct RastaConfigTLS *tls_config;
-        enum RastaTLSMode activeMode;
-        WOLFSSL_CTX *ctx;
-        int file_descriptor;
-        WOLFSSL *ssl;
-        enum RastaTLSConnectionState tls_state;
-    };
-#endif
 
     /**
      * This function will initialise an udp socket and return its file descriptor, which is used to reference it in later
